requirements.txt
Flask
Flask-SQLAlchemy
sentence-transformers
torch
gunicorn
psycopg2-binary
python-dotenv
